+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
  - Fixed `glBufferData` being called to invalidate a buffer created with `glBufferStorage`.
- - Changed use of `Range` in buffer slice to `RangeArgument`.  You will need to import the trait `glium::buffer::RangeArgument` now.
+ - Changed use of `Range` in buffer slice to `RangeArgument`.
 
 ## Version 0.10.0 (2015-10-14)
